11_deployment_implementation/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

flask
gunicorn
boto3
botocore
Pillow
numpy
torch
opencv-python-headless
